+++ cassandra/cqlengine/columns.py
@@ -37,7 +37,7 @@ class BaseValueManager(object):
 
     @property
     def deleted(self):
-        return self.column._val_is_null(self.value) and (self.explicit or self.previous_value is not None)
+        return self.column._val_is_null(self.value) and (self.explicit or not self.column._val_is_null(self.previous_value))
 
     @property
     def changed(self):
@@ -47,7 +47,19 @@ class BaseValueManager(object):
         :rtype: boolean
 
         """
-        return self.value != self.previous_value
+        if self.explicit:
+            return self.value != self.previous_value
+
+        if isinstance(self.column, BaseContainerColumn):
+            default_value = self.column.get_default()
+            if self.column._val_is_null(default_value):
+                return not self.column._val_is_null(self.value) and self.value != self.previous_value
+            elif self.previous_value is None:
+                return self.value != default_value
+
+            return self.value != self.previous_value
+
+        return False
 
     def reset_previous_value(self):
         self.previous_value = deepcopy(self.value)
@@ -57,6 +69,7 @@ class BaseValueManager(object):
 
     def setval(self, val):
         self.value = val
+        self.explicit = True
 
     def delval(self):
         self.value = None