+++ cassandra/cqlengine/connection.py
@@ -17,7 +17,7 @@ import logging
 import six
 import threading
 
-from cassandra.cluster import Cluster, _NOT_SET, NoHostAvailable, UserTypeDoesNotExist
+from cassandra.cluster import Cluster, _NOT_SET, NoHostAvailable, UserTypeDoesNotExist, ConsistencyLevel
 from cassandra.query import SimpleStatement, dict_factory
 
 from cassandra.cqlengine import CQLEngineException
@@ -184,6 +184,8 @@ def register_connection(name, hosts=None, consistency=None, lazy_connect=False,
         conn = Connection.from_session(name, session=session)
         conn.setup_session()
     else:  # use hosts argument
+        if consistency is None:
+            consistency = ConsistencyLevel.LOCAL_ONE
         conn = Connection(
             name, hosts=hosts,
             consistency=consistency, lazy_connect=lazy_connect,